lesson-4-agentdeployment/hosted-agent/requirements.txt
# Azure AI Agent Server Framework - Required for hosted agent
azure-ai-agentserver-agentframework==1.0.0b3

# Microsoft Agent Framework - Core framework for multi-agent orchestration
agent-framework>=0.1.0

# Azure AI Projects SDK - For creating agents and tools
azure-ai-projects>=1.0.0b1

# Azure Identity - Authentication
azure-identity>=1.15.0

# Optional: OpenTelemetry for observability
azure-monitor-opentelemetry>=1.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
